+++ taotie/consumer/base.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, List, Optional
 
 from taotie.storage.base import Storage
-from taotie.utils import Logger
+from taotie.utils.utils import Logger
 
 
 class Consumer(ABC):
