+++ taotie/sources/arxiv.py
@@ -9,7 +9,7 @@ from bs4 import BeautifulSoup
 from taotie.entity import Information
 from taotie.message_queue import MessageQueue, SimpleMessageQueue
 from taotie.sources.base import BaseSource
-from taotie.utils import get_datetime
+from taotie.utils.utils import get_datetime
 
 
 class Arxiv(BaseSource):
