+++ taotie/sources/github.py
@@ -6,7 +6,7 @@ from bs4 import BeautifulSoup
 from taotie.entity import Information
 from taotie.message_queue import MessageQueue
 from taotie.sources.base import BaseSource
-from taotie.utils import *
+from taotie.utils.utils import *
 
 
 class GithubTrends(BaseSource):
